Improve conditional required story
