+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+- If both `.required` and `.optional` are present, the last one to be called takes precedence
+
 ## [1.3.0] - 2019-05-24
 
 ### Changed
